mock data to json

+++ src/models/content.rs
@@ -6,6 +6,6 @@ use fake::{Dummy, Fake, Faker};
 #[derive(Serialize, Deserialize, Debug, Dummy)]
 /// Struct containing image URL
 pub struct Image {
-    pub id: i64,
+    pub id: u32,
     pub path: String,
 }
